+++ tests/test_helpers/utils.py
@@ -84,7 +84,7 @@ def skip_if_no_together(func):
 
 
 def skip_if_no_azureai(func):
-    return pytest.mark.api(skip_if_env_var("AZURE_API_KEY", exists=False)(func))
+    return pytest.mark.api(skip_if_env_var("AZUREAI_API_KEY", exists=False)(func))
 
 
 def skip_if_no_llama_cpp_python(func):